home-lan/osp-raspberrypi-ms/requirements.txt
argparse
numpy<2
opencv-python~=4.5.3.56
tflite-support>=0.4.2
protobuf>=3.18.0,<4
python-socketio[client]==5.8.0
python-engineio==4.7.1